Add warning about minimisation method
We are using the "Powell method", but I'm not sure what the best one
would actually be (when we say best, what do we actually mean? best with
regards to fit results, or most mathematically sound?). The current
choice is a suggestion by ChatGPT, but worth exploring in more detail
once we have everything else sorted.